==> code/3-Report.Rmd <==
---
title: "Assistance tracking in RAIS - Iraq-Jordan-Lebanon"
author: "Initial Dataset Exploration"
date: "Amman, `r Sys.Date()`"
output: 
  word_document: 
    fig_caption: yes
    toc: yes
    toc_depth: 2
    reference_docx: style.docx
---

```{r setup, include=FALSE, echo=FALSE, warning=FALSE, message=FALSE}

source("C:/R-project/assistance-analysis/code/0-packages.R")

assistance.case <- fread( file = "C:/R-project/assistance-analysis/data/assistancecase.csv",na="")
## Reparse correctly the date
assistance.case$ProvidedDate <- as.Date(as.character(assistance.case$ProvidedDate), "%Y-%m-%d")
## need to remove records with wrong date..
assistance.case <- assistance.case[ ProvidedDate > as.Date(fast_strptime("2012-06-01", "%Y-%m-%d"))]
assistance.case <- assistance.case[ ProvidedDate < as.Date(fast_strptime("2017-04-01", "%Y-%m-%d"))]

```



# Introduction

This reports presents the analysis of assistance recorded in the RAIS - (Refugee Assistance Information System) for Jordan, Lebanon & Iraq.

Once cleaned from the records with erroneous date, it represents a total of `r formatC(nrow(assistance.case), big.mark=",")` :   `r formatC(nrow(assistance.case[ instance=="Iraq" ]), big.mark=",")`  for Iraq, `r formatC(nrow(assistance.case[ instance=="Lebanon" ]), big.mark=",")`  for Lebanon and  `r formatC(nrow(assistance.case [ instance == "Jordan" ]), big.mark=",")`  for Jordan.



Assistance Type were recategorised in order to come with meaningfull categories for analysis.

# Analysis of Assistance 

## Assistance per type

```{r assist-type , echo=FALSE, warning=FALSE, fig.width=7}

## Checking structure
#str(assistance.case)
#levels(assistance.case$AssistanceType) 

##frequ <- table (assistance.case[ , c("AssistanceType")])
## with data.table
frequ <- assistance.case[, .N ,by = AssistanceType]
frequ <- frequ[order(frequ$N),]
frequ.level <- as.character(frequ$AssistanceType)
frequ$AssistanceType1 <- factor(frequ$AssistanceType, levels=frequ.level)
#levels(frequ$AssistanceType1)
frequ$Frequency <- frequ$N/sum(frequ$N)

## Old 
# assistance.case$AssistanceType <- factor(assistance.case$AssistanceType, levels=names(frequ[rev(order(frequ, decreasing = TRUE))]))
#assistance.case[ , "AssistanceType"] <- factor(assistance.case[ , "AssistanceType"], levels=names(frequ[rev(order(frequ, decreasing = TRUE))]))

## Because of the size of the dataset, attributes are set differently
## change the levels by reference with no copy within data.table
setattr(assistance.case$AssistanceType,"levels",frequ.level )



## and now the graph
plotfreq <- ggplot(frequ, aes(x=AssistanceType1,y = Frequency)) +
  geom_bar(stat="identity", fill="skyblue3",colour="#2a87c8") +
  #facet_wrap(~ instance, ncol=3) +
  guides(fill=FALSE) +
  ylab("Frequency") +
  #scale_y_continuous(labels=percent)+
  xlab("") +
  coord_flip() +
  ggtitle("Assistance Type")+
  theme(plot.title=element_text(face="bold", size=9),
        plot.background = element_rect(fill = "transparent",colour = NA))

plotfreq


```

## Assistance per type and country

```{r assist-instance , echo=FALSE, warning=FALSE, fig.width=7}


###########################################################################################################
## graph per instance

###########################################################################################################

#frequ.instance <- assistance.case[, .N ,by = .(AssistanceType, instance)]

## Getting directly relative percentage per instance
frequ.instance <-assistance.case[, as.data.table(prop.table(table(AssistanceType, instance),2))]

frequ.instance$AssistanceType1 <- factor(frequ.instance$AssistanceType, levels=frequ.level)
#levels(frequ.instance$AssistanceType1)


plotfreq.instance <- ggplot(frequ.instance, aes(x=AssistanceType1,y = N)) +
  geom_bar(stat="identity",  fill="skyblue3",colour="#2a87c8") +
  facet_wrap(~ instance, ncol=3) +
  guides(fill=TRUE) +
  ylab("Frequency") +
  #scale_y_continuous(labels=percent)+
  xlab("") +
  coord_flip() +
  ggtitle("Assistance Type per instance")+
  theme(plot.title=element_text(face="bold", size=9),
        plot.background = element_rect(fill = "transparent",colour = NA))

plotfreq.instance

```

## Assistance per type, instance & providing organisation

```{r assistprovide , echo=FALSE, warning=FALSE, fig.width=10}

###########################################################################################################
## graph per instance & organisation Provide

###########################################################################################################

#frequ.instance <- assistance.case[, .N ,by = .(AssistanceType, instance)]

## Getting directly relative percentage per instance
frequ.instance.provide <-assistance.case[, as.data.table(prop.table(table(AssistanceType, instance, Provide),2))]
frequ.instance.provide$AssistanceType1 <- factor(frequ.instance.provide$AssistanceType, levels=frequ.level)

## let's subset values that are above 5%
frequ.instance.provide1 <- frequ.instance.provide[frequ.instance.provide$N >= 0.05, ]
frequ.instance.provide1$Provide1 <- as.factor(as.character(frequ.instance.provide1$Provide))

plotfreq.instance.provide <- ggplot(frequ.instance.provide1, aes(x=AssistanceType1,y = N,  fill=Provide)) +
  geom_bar(stat="identity" ,colour="#2a87c8") +
  facet_wrap(~ instance, ncol=3) +
  ylab("Frequency") +
  scale_fill_brewer(palette = "Paired") + 
  xlab("") +
  coord_flip() +
  ggtitle("Assistance Type per instance for delivering Organisation", 
          subtitle = "Only organisation that have assistance records accounting for more than 5% within countries are displayed" )+
  theme(plot.title=element_text(face="bold", size=9),
        plot.subtitle=element_text( size=8),
        legend.position="bottom",
        plot.background = element_rect(fill = "transparent",colour = NA))
plotfreq.instance.provide
```

## Assistance per type, instance & intermediary organisation

```{r assist-through , echo=FALSE, warning=FALSE, fig.width=10}

###########################################################################################################
## graph per instance & organisation Through

###########################################################################################################

#frequ.instance <- assistance.case[, .N ,by = .(AssistanceType, instance)]

## Getting directly relative percentage per instance
frequ.instance.Through <-assistance.case[, as.data.table(prop.table(table(AssistanceType, instance, Through),2))]
frequ.instance.Through$AssistanceType1 <- factor(frequ.instance.Through$AssistanceType, levels=frequ.level)

## let's subset values that are above 2%
frequ.instance.Through1 <- frequ.instance.Through[frequ.instance.Through$N >= 0.01, ]
frequ.instance.Through1$Through1 <- as.factor(as.character(frequ.instance.Through1$Through))

plotfreq.instance.Through <- ggplot(frequ.instance.Through1, aes(x=AssistanceType1,y = N,  fill=Through)) +
  geom_bar(stat="identity" ,colour="#2a87c8") +
  facet_wrap(~ instance, ncol=3) +
  ylab("Frequency") +
  scale_fill_brewer(palette = "Paired") + 
  xlab("") +
  coord_flip() +
  ggtitle("Assistance Type per instance for Intermediate Organisation", 
          subtitle = "Only organisation that have assistance records accounting for more than 1% within countries are displayed" )+
  theme(plot.title=element_text(face="bold", size=9),
        plot.subtitle=element_text( size=8),
        legend.position="bottom",
        plot.background = element_rect(fill = "transparent",colour = NA))
plotfreq.instance.Through
```

## Assistance per type, instance & funding organisation

```{r assist-funding, echo=FALSE, warning=FALSE, fig.width=10}


###########################################################################################################
## graph per instance & organisation Funded

###########################################################################################################

#frequ.instance <- assistance.case[, .N ,by = .(AssistanceType, instance)]

## Getting directly relative percentage per instance
frequ.instance.Funded <-assistance.case[, as.data.table(prop.table(table(AssistanceType, instance, Funded),2))]
frequ.instance.Funded$AssistanceType1 <- factor(frequ.instance.Funded$AssistanceType, levels=frequ.level)

## let's subset values that are above 1%
frequ.instance.Funded1 <- frequ.instance.Funded[frequ.instance.Funded$N >= 0.01, ]
frequ.instance.Funded1$Funded1 <- as.factor(as.character(frequ.instance.Funded1$Funded))

plotfreq.instance.Funded <- ggplot(frequ.instance.Funded1, aes(x=AssistanceType1,y = N,  fill=Funded)) +
  geom_bar(stat="identity" ,colour="#2a87c8") +
  facet_wrap(~ instance, ncol=3) +
  ylab("Frequency") +
  scale_fill_brewer(palette = "Paired") + 
  xlab("") +
  coord_flip() +
  ggtitle("Assistance Type per instance for Funding Organisation", 
          subtitle = "Only organisation that have assistance records accounting for more than 1% within countries are displayed" )+
  theme(plot.title=element_text(face="bold", size=9),
        plot.subtitle=element_text( size=8),
        legend.position="bottom",
        plot.background = element_rect(fill = "transparent",colour = NA))
plotfreq.instance.Funded
```

# Analysis per time

## Assistance per type, instance & date

```{r assist-time , echo=FALSE, warning=FALSE, fig.width=15, fig.height=8}

###########################################################################################################
## graph per instance & date

###########################################################################################################


#frequ.instance.date <-assistance.case[, as.data.table(table(AssistanceType, instance, year(ProvidedDate)))]
#frequ.instance.month <-assistance.case[, as.data.table(table(AssistanceType, instance, month(ProvidedDate)))]
frequ.instance.monthyear <-assistance.case[, as.data.table(table(AssistanceType, instance, format(ProvidedDate, "%Y-%m")))]
 
## Reorder factors
frequ.instance.monthyear$AssistanceType1 <- factor(frequ.instance.monthyear$AssistanceType, levels=frequ.level)

plotfreq.instance.monthyear <- ggplot(frequ.instance.monthyear, aes(x=V1, y = N,  fill=AssistanceType1)) +
  geom_bar(stat="identity" ,colour="#2a87c8") +
  facet_wrap(~ instance, nrow=3) +
  ylab("# of records") +
  scale_fill_brewer(palette = "Paired") + 
  xlab("") +
  
  ## setting up a focus in 
 # coord_flip() +
  ggtitle("Assistance Type per instance and year", 
          subtitle = " " )+
  theme(plot.title=element_text(face="bold", size=9),
        plot.subtitle=element_text( size=8),
        plot.background = element_rect(fill = "transparent",colour = NA),
        #legend.position="top",
        legend.position="bottom",
        legend.title=element_blank(),
        axis.text.x = element_text(angle = 30, vjust = 0.5, hjust=1))
plotfreq.instance.monthyear

```


# Analysis per subtype

## Details of cash assistance

## Details of NFI assistance

# Analysis of assistance flows

## Assistance Type from & through - by year & country
The following sankey diagram provided an overview of the origin and delivery stream of 


# Assistance occurence by case

The following visualisation provides an overview of the most associated assistance type
